move apollo files into apollo directory
