+++ packages/explorer-2.0/lib/utils.tsx
@@ -493,7 +493,7 @@ export function getTimeframe(timeWindow) {
 }
 
 /**
- * gets the amoutn difference plus the % change in change itself (second order change)
+ * gets the amount difference plus the % change in change itself (second order change)
  * @param {*} valueNow
  * @param {*} value24HoursAgo
  * @param {*} value48HoursAgo
