+++ packages/explorer-2.0/pages/index.tsx
@@ -3,7 +3,7 @@ import { useQuery } from '@apollo/client'
 import Orchestrators from '../components/Orchestrators'
 import { useWeb3React } from '@web3-react/core'
 import { getLayout } from '../layouts/main'
-import { withApollo } from '../lib/apollo'
+import { withApollo } from '../apollo'
 import { Box } from 'theme-ui'
 import Approve from '../components/Approve'
 import Search from '../components/Search'
