+++ packages/explorer-2.0/pages/voting/[poll].tsx
@@ -5,13 +5,10 @@ import fm from 'front-matter'
 import IPFS from 'ipfs-mini'
 import { Box } from 'theme-ui'
 import Card from '../../components/Card'
-import ReactTooltip from 'react-tooltip'
-import Help from '../../public/img/help.svg'
 import VotingWidget from '../../components/VotingWidget'
 import ReactMarkdown from 'react-markdown'
-import Utils from 'web3-utils'
 import { abbreviateNumber } from '../../lib/utils'
-import { withApollo } from '../../lib/apollo'
+import { withApollo } from '../../apollo'
 import { useRouter } from 'next/router'
 import { useQuery, useApolloClient, gql } from '@apollo/client'
 import { useWeb3React } from '@web3-react/core'
@@ -238,24 +235,6 @@ const Poll = () => {
                     <Box sx={{ color: 'muted' }}>
                       Total Support ({pollData.quota / 10000}% needed)
                     </Box>
-                    {/* <Flex>
-                      <ReactTooltip
-                        id="tooltip-total-staked"
-                        className="tooltip"
-                        place="top"
-                        type="dark"
-                        effect="solid"
-                      />
-                      <Help
-                        data-tip="This is the amount currently delegated to an Orchestrator."
-                        data-for="tooltip-total-staked"
-                        sx={{
-                          color: 'muted',
-                          cursor: 'pointer',
-                          ml: 1,
-                        }}
-                      />
-                    </Flex> */}
                   </Flex>
                 }
                 subtitle={
@@ -320,24 +299,6 @@ const Poll = () => {
                     <Box sx={{ color: 'muted' }}>
                       Total Participation ({pollData.quorum / 10000}% needed)
                     </Box>
-                    {/* <Flex>
-                      <ReactTooltip
-                        id="tooltip-equity"
-                        className="tooltip"
-                        place="top"
-                        type="dark"
-                        effect="solid"
-                      />
-                      <Help
-                        data-tip="Account's equity relative to the entire network."
-                        data-for="tooltip-equity"
-                        sx={{
-                          color: 'muted',
-                          cursor: 'pointer',
-                          ml: 1,
-                        }}
-                      />
-                    </Flex> */}
                   </Flex>
                 }
                 subtitle={
