+++ packages/explorer-2.0/pages/voting/create-poll.tsx
@@ -11,7 +11,7 @@ import gql from 'graphql-tag'
 import { useWeb3React } from '@web3-react/core'
 import PollTokenApproval from '../../components/PollTokenApproval'
 import { useQuery } from '@apollo/client'
-import { withApollo } from '../../lib/apollo'
+import { withApollo } from '../../apollo'
 import { useApolloClient } from '@apollo/client'
 import { MutationsContext } from '../../contexts'
 import Utils from 'web3-utils'
