+++ packages/explorer-2.0/pages/voting/index.tsx
@@ -3,7 +3,7 @@ import { Text } from '@theme-ui/components'
 import { getLayout } from '../../layouts/main'
 import { Box } from 'theme-ui'
 import Button from '../../components/Button'
-import { withApollo } from '../../lib/apollo'
+import { withApollo } from '../../apollo'
 import { useQuery } from '@apollo/client'
 import Spinner from '../../components/Spinner'
 import Card from '../../components/Card'
